feat#756: add prod user-tiers

+++ src/product-watchlist-api/tests/watchlist_post.rs
@@ -59,8 +59,7 @@ async fn should_201_when_new_watchlist_entry_would_not_exceed_quota() {
     let service =
         ProductWatchListServiceImpl::new(&watchlist_repository, &product_repository, &user_service);
 
-    let product_records =
-        fake::vec![ProductRecord; user::core::tier::UserTier::Free.watchlist_limit() - 1];
+    let product_records = fake::vec![ProductRecord; (user::core::tier::UserTier::Free.watchlist_limit() - 1) as usize];
     let put_res = product_repository
         .put_product_records(product_records.clone().try_into().unwrap())
         .await
@@ -162,7 +161,7 @@ async fn should_422_when_new_watchlist_entry_would_exceed_quota() {
         ProductWatchListServiceImpl::new(&watchlist_repository, &product_repository, &user_service);
 
     let product_records =
-        fake::vec![ProductRecord; user::core::tier::UserTier::Free.watchlist_limit()];
+        fake::vec![ProductRecord; user::core::tier::UserTier::Free.watchlist_limit() as usize];
     let put_res = product_repository
         .put_product_records(product_records.clone().try_into().unwrap())
         .await
